move 'fromJSON' parsing code from Attachment into AttachmentParser
