+++ app/src/main/java/co/ello/android/ello/model/Attachment.kt
@@ -15,16 +15,4 @@ data class Attachment(
     var image: Image? = null
 
     val isGif: Boolean get() = type == "image/gif" || url.getPath().endsWith(".gif")
-
-    companion object {
-        fun fromJSON(json: JSON): Attachment {
-            val url = json["url"].url!!
-            val attachment = Attachment(url)
-            attachment.size = json["metadata"]["size"].int
-            attachment.width = json["metadata"]["width"].int
-            attachment.height = json["metadata"]["height"].int
-            attachment.type = json["metadata"]["type"].string
-            return attachment
-        }
-    }
 }
